// src/app/Models/contact.ts
export interface Contact{
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    profileImage?: string;
}


// export interface Contact {
//     id: number;
//     firstName: string;
//     lastName: string;
//     email: string;
//     phone: string;
//     userId: string;  
//   }
  
  